Removed some unused crud from QWDOS.  Added non-move sbrk fix since QWDOS can work in Win9x dosbox with NDIS3PK2 Fixed some unresolved stuff at link time for Win32 QWCL

+++ QW/R_MISC.C
@@ -203,7 +203,7 @@ void R_NetGraph (void)
 	else
 		w = NET_TIMINGS;
 
-	x =	-((vid.width - 320)>>1);
+	x =	-(( (signed int)vid.width - 320)>>1);
 	y = vid.height - sb_lines - 24 - (int)r_graphheight.value*2 - 2;
 
 	M_DrawTextBox (x, y, (w+7)/8, ((int)r_graphheight.value*2+7)/8 + 1);
